refactor: 승패 결과를 판단하는 책임을 Dealer로 이동
