+++ src/main/java/domain/GameResult.java
@@ -1,8 +1,5 @@
 package domain;
 
-import domain.participant.Dealer;
-import domain.participant.Player;
-
 public enum GameResult {
     WIN,
     LOSE,
@@ -15,18 +12,4 @@ public enum GameResult {
             case DRAW -> DRAW;
         };
     }
-
-    public static GameResult getResult(Player player, Dealer dealer) {
-        if (player.isBurst()) {
-            return GameResult.LOSE;
-        }
-        if (dealer.isBurst() || player.calculateScore() > dealer.calculateScore()) {
-            return GameResult.WIN;
-        }
-        if (player.calculateScore() < dealer.calculateScore()) {
-            return GameResult.LOSE;
-        }
-        return GameResult.DRAW;
-    }
-
 }
